requirements.txt
aiogram
yookassa
loguru
environs
aiohttp
configparser
python-dotenv
groq